change optional prop declerations; migrate stories to TS and fix issues
