+++ src/components/Buttons/Button/Button.tsx
@@ -7,19 +7,19 @@ import { ButtonType, Context, Size } from '../../../constants';
 // supported although not defined in propTypes
 interface Props extends Omit<React.ButtonHTMLAttributes<HTMLButtonElement>, 'size'> {
     /** The label of the button */
-    children: React.ReactElement;
+    children: string | React.ReactElement;
     /** The button context (e.g. brand, primary, bad, good etc. - defaults to neutral) */
-    context: Context | 'link';
+    context?: Context | 'link';
     /** The size of the button */
-    size: Size;
+    size?: Size;
     /** Whether or not to show block-level button (full width) */
-    isBlock: boolean;
+    isBlock?: boolean;
     /** Whether or not to show inline button (without padding) */
-    isInline: boolean;
+    isInline?: boolean;
     /** Should button be disabled or not */
-    disabled: boolean;
+    disabled?: boolean;
     /** Type of the button */
-    type: ButtonType;
+    type?: ButtonType;
     /** Providing an href will render an <a> element, styled as a button. */
     href?: string;
 }
