+++ src/components/Callout/Callout.tsx
@@ -6,13 +6,13 @@ import styles from './Callout.scss';
 
 interface Props {
     /** The Callout context (e.g. brand, primary, bad, good etc. - defaults to info) */
-    context: Context;
+    context?: Context;
     /** A property specifies the height of a line. */
-    lineHeightStyle: number;
+    lineHeightStyle?: number;
     /** Content to be rendered inside the container */
-    children: React.ReactElement;
+    children: string | React.ReactElement;
     /** A function to be called when close button was clicked */
-    onRequestClose: () => void;
+    onRequestClose?: () => void;
 }
 
 const { block, elem } = bem('Callout', styles);
