+++ src/components/Icon/IconTextkernel/IconTextkernel.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react';
+import IconBase, { Props as BaseProps } from '../IconBase';
+
+interface Props extends Omit<Omit<BaseProps, 'children'>, 'viewbox'> {}
+
+const IconTextkernel: React.FC<Props> = props => (
+    <IconBase {...props} viewBox="0 0 369.7 800">
+        <path d="M334.3,800C203.4,766.9,89.9,681.8,89.9,543.8V248.3H0v-104H89.9V0H223.1V144.2H349.2v104H223.1V554c0,76.5,56.7,109.6,146.6,130.8Z" />
+    </IconBase>
+);
+
+IconTextkernel.displayName = 'IconTextkernel';
+
+export default IconTextkernel;
