+++ src/components/Navigation/NavItem/NavItem.tsx
@@ -4,11 +4,11 @@ import styles from './NavItem.scss';
 
 interface Props extends React.HTMLAttributes<HTMLDivElement> {
     /** Add active class to the item. For controlled components */
-    active: boolean;
+    active?: boolean;
     /** If set to true, this item will be pulled to the right. Only one NavItem inside a nav bar should have this attribute */
-    pullRight: boolean;
+    pullRight?: boolean;
     /** add activeClassName to the rendered item. You should add this to react-routers NavLink */
-    useActiveClass: boolean;
+    useActiveClass?: boolean;
     /** Node to be rendered as a single navigation item. There should be only 1 child */
     children: React.ReactElement;
 }
