+++ stories/WidthRestrictor.tsx
